✅ 824XT0 code: done
Task: 202604210859-824XT0
Primary: code
Agent: CODER
Author: CODER
Status: DONE
Comment: verified: split approved oversized recipe and query-run-execute tests by scenario family | details: split files are below 600 lines and affected cli-recipes/cli-core tests, lint, diff check, policy routing, and framework bootstrap passed.
